moved commands to separate system
